back/requirements.txt
boto3==1.20.28
botocore==1.23.28
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
mysqlclient==2.1.0
PyJWT==2.3.0
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.29
urllib3==1.26.7
Werkzeug==2.0.2
